simple authorize

+++ cookies/admin.php
@@ -1,14 +1,14 @@
 <?php
-    include("vendor/autoload.php");
+include("vendor/autoload.php");
 
-    use Libs\Database\MySQL;
-    use Libs\Database\UsersTable;
-    use Helpers\Auth;
+use Libs\Database\MySQL;
+use Libs\Database\UsersTable;
+use Helpers\Auth;
 
-    $auth = Auth::check();
+$auth = Auth::check();
 
-    $table = new UsersTable(new MySQL);
-    $users = $table->fetchAll();
+$table = new UsersTable(new MySQL);
+$users = $table->fetchAll();
 ?>
 <!DOCTYPE html>
 <html lang="en">
@@ -22,6 +22,7 @@
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.8/dist/js/bootstrap.bundle.min.js" defer></script>
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.4.0/css/all.min.css">
 </head>
+
 <body>
     <nav class="navbar navbar-dark bg-dark navbar-expand">
         <div class="container">
@@ -41,7 +42,7 @@
             </ul>
         </div>
     </nav>
-<div class="container mt-4">
+    <div class="container mt-4">
         <table class="table table-bordered table-striped">
             <tr>
                 <th>ID</th>
@@ -51,7 +52,7 @@
                 <th>Role</th>
                 <th></th>
             </tr>
-            <?php foreach($users as $user): ?>
+            <?php foreach ($users as $user): ?>
                 <tr>
                     <td><?= $user->id ?></td>
                     <td><?= $user->name ?></td>
@@ -59,41 +60,47 @@
                     <td><?= $user->phone ?></td>
                     <td>
                         <div class="d-flex justify-content-between align-items-center">
-                            <?php if($user->role_id == 3): ?>
-                            <span class="badge bg-success">
-                                <?= $user->role ?>
-                            </span>
-                        <?php elseif($user->role_id == 2): ?>
-                            <span class="badge bg-primary">
-                                <?= $user->role ?>
-                            </span>
-                        <?php else: ?>
-                            <span class="badge bg-secondary">
-                                <?= $user->role ?>
-                            </span>
-                        <?php endif ?>
+                            <?php if ($user->role_id == 3): ?>
+                                <span class="badge bg-success">
+                                    <?= $user->role ?>
+                                </span>
+                            <?php elseif ($user->role_id == 2): ?>
+                                <span class="badge bg-primary">
+                                    <?= $user->role ?>
+                                </span>
+                            <?php else: ?>
+                                <span class="badge bg-secondary">
+                                    <?= $user->role ?>
+                                </span>
+                            <?php endif ?>
 
-                        <div class="btn-group dropdown">
-                            <a href="#" class="btn btn-sm dropdown-toggle" data-bs-toggle="dropdown"></a>
-                            <div class="dropdown-menu dropdown-menu-dark dropdown-menu-end">
-                                <a href="_actions/role.php?id=<?= $user->id ?>&role=1" class="dropdown-item">User</a>
-                                <a href="_actions/role.php?id=<?= $user->id ?>&role=2" class="dropdown-item">Manager</a>
-                                <a href="_actions/role.php?id=<?= $user->id ?>&role=3" class="dropdown-item">Admin</a>
-                            </div>
-                        </div>
+                            <?php if ($auth->role_id == 3): ?>
+                                <div class="btn-group dropdown">
+                                    <a href="#" class="btn btn-sm dropdown-toggle" data-bs-toggle="dropdown"></a>
+                                    <div class="dropdown-menu dropdown-menu-dark dropdown-menu-end">
+                                        <a href="_actions/role.php?id=<?= $user->id ?>&role=1" class="dropdown-item">User</a>
+                                        <a href="_actions/role.php?id=<?= $user->id ?>&role=2" class="dropdown-item">Manager</a>
+                                        <a href="_actions/role.php?id=<?= $user->id ?>&role=3" class="dropdown-item">Admin</a>
+                                    </div>
+                                </div>
+                            <?php endif ?>
                         </div>
                     </td>
                     <td>
                         <div class="d-flex justify-content-between">
-                            <a href="_actions/delete.php?id=<?= $user->id ?>" class="btn btn-sm btn-outline-danger">Delete</a>
-                            <?php if($user->suspended): ?>
-                                <a href="_actions/unsuspend.php?id=<?= $user->id ?>" style="color: grey">
-                                    <i class="fa-solid fa-lock"></i>
-                                </a>
-                            <?php else: ?>
-                                <a href="_actions/suspend.php?id=<?= $user->id ?>" class="btn btn-sm btn-outline-warning">
-                                    Ban
-                                </a>
+                            <?php if ($auth->role_id == 3): ?>
+                                <a href="_actions/delete.php?id=<?= $user->id ?>" class="btn btn-sm btn-outline-danger">Delete</a>
+                            <?php endif ?>
+                            <?php if ($auth->role_id >= 2): ?>
+                                <?php if ($user->suspended): ?>
+                                    <a href="_actions/unsuspend.php?id=<?= $user->id ?>" style="color: grey">
+                                        <i class="fa-solid fa-lock"></i>
+                                    </a>
+                                <?php else: ?>
+                                    <a href="_actions/suspend.php?id=<?= $user->id ?>" class="btn btn-sm btn-outline-warning">
+                                        Ban
+                                    </a>
+                                <?php endif ?>
                             <?php endif ?>
                         </div>
                     </td>
